modifier la fonction de reduction qui prenait pas en compte les 0 et modif géneration de mdp en évitant les doublons

+++ genererMdp.php
@@ -16,30 +16,30 @@
    //etape 2 parcourir le tableau et prendre 6lettre et 4chiffre
     $nblettre=0;
     $nbchiffre=0;
-    $nouveauMdp;
+    $nouveauMdp=null;
         foreach ($tableauCaractere as $key => $caractere) {
-            if($nbchiffre<4||$nblettre<6){
-                if (is_numeric($caractere)) {
+                if (is_numeric($caractere)&&strlen($nouveauMdp)<10) {
                     if($nbchiffre<4){
                         $nbchiffre++;
-                        if(empty($nouveauMdp)){
+                        if(!isset($nouveauMdp)){
                             $nouveauMdp=$caractere;
                         }else {
                             $nouveauMdp.=$caractere;
                         }
                     }
                 }else {
-                    if($nblettre<6){
+                    if($nblettre<6&&strlen($nouveauMdp)<10){
                         $nblettre++;
-                        if(empty($nouveauMdp)){
+                        if(!isset($nouveauMdp)){
                             $nouveauMdp=$caractere;
                         }else {
                             $nouveauMdp.=$caractere;
                         }
                     }
-                }
-            }
+                } 
+            
         }
+    
    //etape 3 retourner le hash reduit
    return $nouveauMdp;
  }
@@ -108,7 +108,7 @@
     //$handle=fopen("mdpFile.txt", "w+");
     //fclose($handle);
     //le nb de boucle à parcourir essayer avec 10 moi sa bug
-    for ($i=0; $i <10 ; $i++) {
+   // for ($i=0; $i <10 ; $i++) {
         //etape 1 ouvrir le fichier et récuperer les mdp 
         $lesMdp = file_get_contents('mdpFile.txt');
         $lesMdp=explode("\r\n",$lesMdp);
@@ -122,9 +122,14 @@
                     echo "</br> le hasher: ".$hash."</br>";
                     $mdp=reduce($hash);
                     echo "</br> le mdp equivalent: ".$mdp."</br>";
-                    //on écrit chaque mdp dans le fichier mdpFile.txt
-                    file_put_contents("mdpFile.txt",$mdp."\r\n", FILE_APPEND);
 
+                    $mdps = file_get_contents("mdpFile.txt"); //read names into string
+                    if(false === strpos($mdps,$mdp)) { //write name if it's not there already
+                        //on écrit chaque mdp dans le fichier mdpFile.txt
+                        file_put_contents("mdpFile.txt",$mdp."\r\n", FILE_APPEND);
+                    }
+                    
+                    
 
                 }
             }
@@ -135,13 +140,18 @@
                 echo "</br> le hash: ".$hash."</br>";
                 $mdp=reduce($hash);
                 echo "</br> le mdp equivalent: ".$mdp."</br>";
-                //on écrit chaque mdp dans le fichier mdpFile.txt
-                file_put_contents("mdpFile.txt",$mdp."\r\n", FILE_APPEND);
+
+                
+                $mdps = file_get_contents("mdpFile.txt"); //read names into string
+                if(false === strpos($mdps,$mdp)) { //write name if it's not there already
+                    //on écrit chaque mdp dans le fichier mdpFile.txt
+                    file_put_contents("mdpFile.txt",$mdp."\r\n", FILE_APPEND);
+                }
 
             }
         
         }
-    }
+    //}
    
   
   }
@@ -237,5 +247,12 @@
                               "g5f6s3t4gv",
                               "bf896dez5z",
 "xvdq5ze874"));*/
+
+/*laisser le meme nombre d'iteration
+*et actualiser la page plusieur fois 
+*et voir que le nb de mdp augmente
+*/
+for ($i=0; $i <2 ; $i++) { 
     genererMdp("bb8327ceab");
+}
 ?>
